Parser: Heavily rework function call & field access type checking
Due to the grammar changes, the type checking for function calls
and field accesses can be done the same way. The novelty is that
function calls and field accesses are treated as objects of types
that allow certain functions or fields. E.g., an OpenAPI component
allows for get, put, etc. in combination with routes. These new
types can seamlessly blend in with the attribute grammar used for
expressions.