Parser: Major rework of the type checking Merge walkers into one to improve efficiency. Instead of walking the tree several times, we walk it only once.
To measure this, we added a simple class that collects measurements.

Introduce scopes where necessary (e.g., blocks) and make sure that
identifiers are not re-used.

Types are now being treated as singletons. This is trivial for
primitive types, but can get a bit more difficult for complex types.
Hence, there now is a distinction between primitive and complex types.
It is up to the complex type whether it was already registered in
the `TypeRegistry`, i.e., it depends on its relevant fields.
The OpenAPI type for example is "uniquely" identified by being an
OpenAPIComponent instance AND its URL.