Parser: More refactoring of the type system
This is a follow-up of the last type system change. It now supports
'type caching'. This is especially helpful when used with
a language server. ANTLR has a global cache which already simplifies
a subsequent "diagnosis" calls with regards to lexing and parsing.
The type system caching speeds-up the tree-walking
(aka semantic analysis). This speed-up is significant
when the code includes an OpenAPI component. Due to the caching
the API specification does not have to be fetched (again). A cold
start of the parser (e.g., file opened in editor) takes ~1.5 s
including lexing, parsing, semantic analysis (i.e., a full diagnosis).
A subsequent call with "minor" changes (e.g., adding an else branch
to an existing if statement) takes mere 4 ms for a full diagnosis.